refactor: extracted getNexus
