+++ src/sdk/account/decorators/multichainRead.ts
@@ -1,16 +1,11 @@
 import { readContract } from "viem/actions"
-import { erc7579Reads } from "../../clients/decorators/erc7579"
 import type { ReadDictionary } from "../../clients/decorators/erc7579"
 import type { AnyData, ModularSmartAccount } from "../../modules/utils/Types"
-import { ownableReads } from "../../modules/validators/ownable/decorators"
+import {
+  GLOBAL_COMPOSABLE_READS,
+  type SupportedRead
+} from "../utils/composableReads"
 import type { BaseInstructionsParams } from "./build"
-
-export const GLOBAL_COMPOSABLE_READS = {
-  ...erc7579Reads,
-  ...ownableReads
-} as const
-
-export type SupportedRead = keyof typeof GLOBAL_COMPOSABLE_READS
 
 // biome-ignore lint/complexity/noBannedTypes: Later inference will be used
 type ArgumentTypes<F extends Function> = F extends (
@@ -36,14 +31,12 @@ export const multichainRead = async <T>(
 
   // Cast GLOBAL_COMPOSABLE_READS to ReadDictionary to ensure all read functions have a consistent type
   const readFunctions = GLOBAL_COMPOSABLE_READS as unknown as ReadDictionary
-
   const results = await Promise.all(
     account.deployments.map(async (account) => {
       const chainId = account.client.chain?.id
       if (!chainId) {
         throw new Error("Chain ID is not set")
       }
-
       const [readData] = await readFunctions[type](
         account,
         parametersForType as AnyData
